feat(chat): reconnect chat after page reload or when going back
